agent: extract email infrastructure into agent::email module
Move EmailSender trait, Sender enum, and ResendSender out of agent::alerts::notifier into a standalone agent::email module so non-alert emails (e.g. invite emails) can reuse the same sending infrastructure.

- New agent::email module with EmailSender trait, Sender, ResendSender, and Arc<Sender> impl
- Extract EMAIL_WRAPPER_TEMPLATE constant and register_email_wrapper() from notifications crate so non-alert emails can share Estuary branding
- Construct Sender at top level in main.rs wrapped in Arc, ready to be shared across alert notifications and future invite email code
- alerts::notifier re-exports EmailSender and Sender to preserve the existing public API
