+++ crates/notifications/src/lib.rs
@@ -212,15 +212,9 @@ impl Renderer {
     }
 }
 
-fn register_common_templates<'a>(registry: &mut handlebars::Handlebars<'a>) -> anyhow::Result<()> {
-    // Common email wrapper template that's used by all alert types, so we have consistent styling.
-    // Note that this html was generated by an LLM, from the original `mjml` template in the legacy
-    // alerts edge function. That legacy template used the `mjml-browser` library to render html
-    // with styling that matches our UI. This LLM-translation doesn't match that _perfectly_,
-    // but was considered good enough to allow getting rid of `mjml`.
-    registry.register_template_string(
-        "email_wrapper",
-        r#"<!DOCTYPE html>
+/// Branded HTML wrapper template shared by all Estuary emails (alerts, invites, etc.).
+/// Expects `full_name` (optional) and a `body_template_name` partial to inject content.
+pub const EMAIL_WRAPPER_TEMPLATE: &str = r#"<!DOCTYPE html>
 <html>
 <head>
     <meta charset="UTF-8">
@@ -269,8 +263,18 @@ fn register_common_templates<'a>(registry: &mut handlebars::Handlebars<'a>) -> a
         </div>
     </div>
 </body>
-</html>"#,
-    ).context("registering email_wrapper template")?;
+</html>"#;
+
+/// Registers the shared email wrapper template. Call this before registering
+/// any body templates that use `{{> (lookup this "body_template_name")}}`.
+pub fn register_email_wrapper(registry: &mut handlebars::Handlebars) -> anyhow::Result<()> {
+    registry
+        .register_template_string("email_wrapper", EMAIL_WRAPPER_TEMPLATE)
+        .context("registering email_wrapper template")
+}
+
+fn register_common_templates<'a>(registry: &mut handlebars::Handlebars<'a>) -> anyhow::Result<()> {
+    register_email_wrapper(registry)?;
 
     // Helper partial for rendering a catalog name identifier
     registry
